requirements.txt
numpy
pandas
scikit-learn
scipy
sklearn
tensorflow==1.15
tensorflow-gpu==1.15
pillow
inputs
opencv-python
pywin32
matplotlib